refactored the code and put the various components into own functions and files
